move reporters file
